pyproject.toml
[tool.poetry]
name = "job-portal-graphql"
version = "0.1.0"
description = "A Flask-based job portal application with GraphQL"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = []
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
alembic = "1.15.2"
APScheduler = "3.11.0"
ariadne = "0.26.2"
bcrypt = "4.3.0"
blinker = "1.9.0"
cachetools = "5.5.2"
certifi = "2025.1.31"
charset-normalizer = "3.4.1"
click = "8.1.8"
dnspython = "2.7.0"
email_validator = "2.2.0"
Flask = "3.1.0"
Flask-Bcrypt = "1.0.1"
Flask-Login = "0.6.3"
Flask-Mail = "0.10.0"
Flask-Migrate = "4.1.0"
Flask-SQLAlchemy = "3.1.1"
flask-talisman = "1.1.0"
Flask-WTF = "1.2.2"
google-api-core = "2.24.2"
google-auth = "2.38.0"
google-cloud-core = "2.4.3"
google-cloud-storage = "3.1.0"
google-crc32c = "1.7.1"
google-resumable-media = "2.7.2"
googleapis-common-protos = "1.69.2"
greenlet = "3.1.1"
gunicorn = "23.0.0"
idna = "3.10"
itsdangerous = "2.2.0"
Jinja2 = "3.1.6"
Mako = "1.3.9"
MarkupSafe = "3.0.2"
packaging = "24.2"
pillow = "11.1.0"
proto-plus = "1.26.1"
protobuf = "6.30.2"
psycopg2-binary = "2.9.10"
pyasn1 = "0.6.1"
pyasn1_modules = "0.4.2"
python-dotenv = "1.1.0"
requests = "2.32.3"
rsa = "4.9"
SQLAlchemy = "2.0.40"
typing_extensions = "4.13.1"
tzlocal = "5.3.1"
urllib3 = "2.3.0"
Werkzeug = "3.1.3"
WTForms = "3.2.1"
graphql-core = "3.2.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.1.0"
black = "^23.3.0"
flake8 = "^6.0.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
